feature: add a plugin for faster webserver.
