--- plugins/webserver/configs/__init__.py
@@ -1,3 +1,0 @@
-from .settings import settings
-
-__all__ = ["settings"]
