--- plugins/webserver/configs/webserver_config_input.py
@@ -1,5 +1,0 @@
-from typing_extensions import TypedDict
-
-
-class WebServerConfigInput(TypedDict):
-    stream: bool
