+++ plugins/webserver/consts.py
@@ -1,0 +1,12 @@
+# parquet files generated from indexing pipeline
+COMMUNITY_REPORT_TABLE = "create_final_community_reports"
+ENTITY_TABLE = "create_final_nodes"
+ENTITY_EMBEDDING_TABLE = "create_final_entities"
+
+# community level in the Leiden community hierarchy from which we will load the community reports
+# higher value means we use reports from more fine-grained communities (at the cost of higher computation cost)
+COMMUNITY_LEVEL = 2
+
+RELATIONSHIP_TABLE = "create_final_relationships"
+COVARIATE_TABLE = "create_final_covariates"
+TEXT_UNIT_TABLE = "create_final_text_units"
