--- plugins/webserver/search/__init__.py
@@ -1,2 +1,0 @@
-from .localsearch import build_local_question_gen, build_local_search_engine, load_local_context
-from .globalsearch import build_global_search_engine, build_global_context_builder
