--- plugins/webserver/utils/__init__.py
@@ -1,4 +1,0 @@
-from .files import get_sorted_subdirs
-from .refer import get_reference
-
-__all__ = ["get_sorted_subdirs"]
